+++ pid/pid_linux.go
@@ -10,7 +10,7 @@ package pid
 import (
 	"fmt"
 
-	"pkg.re/essentialkaos/ek.v2/fsutil"
+	"pkg.re/essentialkaos/ek.v3/fsutil"
 )
 
 // ////////////////////////////////////////////////////////////////////////////////// //