+++ changelog.md
@@ -9,6 +9,7 @@
 * `[timeutil]` Code refactoring
 * `[system]` Code refactoring
 * `[log]` Code refactoring
+* `[arg]` Code refactoring
 
 ---