+++ changelog.md
@@ -1,5 +1,18 @@
 ## Changelog
 
+#### v3.0.0
+
+* `[fmtutil]` Pluralization moved from `fmtutil` to separate package `pluralize` (_incompatible changes_)
+* `[pluralize]` Brand new pluralization package with more than 140 languages support
+* `[timeutil]` Improved `PrettyDuration` output
+* `[system]` Now `SessionInfo` contnains full user info (`Info` struct) instead username (_incompatible changes_)
+* `[timeutil]` Code refactoring
+* `[system]` Code refactoring
+* `[log]` Code refactoring
+* `[arg]` Code refactoring
+
+---
+
 #### v2.0.2
 
 * `[pid]` Added function `IsWorks` which return true if process with pid from pidfile is active